created membersGuest.js, moved members styles to separate file
